[tint][val] Improve input_index_attachement validation
This is less of a migration, and instead using the ValidateIOAttribute
framework to test for the attribute being on function params and
returns, which are not Vars. The original Var code is retained to
cover MSVs and other cases like function scoped vars.

This involved loosen when ValidatorIOAttributes runs to handle MSVs
that are not in the In or Out address space. This will also be needed
for dealing with binding_point in the future, which is why I went with
a more systematic change instead of just special casing it.

Bug: 455376684
Change-Id: Ie7fa4cc17d91db3afc6c0a083d0643d603c3b5c9
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/274894
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: James Price <jrprice@google.com>