GL: Add toggle for multi-threaded context current
Adds a toggle to control whether a GL context can be made current on
different threads.

When disabled, the context is assumed to be used on a single thread,
avoiding the overhead of frequent context switches. This is beneficial
for use cases like WebGPU that operate on a single thread.

Bug: 451928481
Change-Id: I33b85dfdb6180f2d570faa2f5920bb31009ceae6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/274094
Reviewed-by: Stephen White <senorblanco@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>