[dawn][native] Fix BlitTextureToBuffer handling of undefined bytesPerRow and rowsPerImage
Make sure to use layout values that are not undefined when calling
BlitTextureToBuffer. This is done via
ApplyDefaultTexelCopyBufferLayoutOptions, which replaces undefined
values with the real values, correctly computed in block space and
aligned. The code in BlitTextureToBuffer was handling these values when
undefined incorrectly. In practice, this means they would do the wrong
thing when bytesPerRow is undefined and the texture is compressed.

Bug: 462167738
Bug: 424536624
Change-Id: I1a78bb5b17ad880b06a6b82d1289781ddaada5b4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/275234
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>