[tint][val] Merge Locations + Blend Src and IOAttributes validation
This CL moves the iterating over various IO objects (params, returns,
and MSVs) into ValidateIOAttributes, so there can be a single callsite
in CheckFunction. The original implementation is moved off into an
impl function ValidateIOAttributesImpl.

The logic for CheckEntryPointLocationsAndBlendSrc is broken up from 2
passes into 3 passes; populate context, validate blend_src, and
validate location.

These passes are moved into the top-level ValidateIOAttributes
function, so that all of the IO attributes validation is invoked via a
single call.

The logic for validating blend_src and location is NOT moved into
IOAttributeCheckers in this CL, because that will require significant
reworking of the existing logic, and grouping all of the IO attributes
logic into a single callsite is an improvement to the code's
readability, even if the migration to Checker does not happen.

Bug: 455376684

Change-Id: I41fe254419952a5369e6594bdfb3ae8f1d452f7f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/275454
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>