vulkan: Initialize buffer padding on CPU
There is special code to eagerly transition buffer usage back to
MapWrite from UploadData(). This is workaround for host visible buffers
so they can use the fast upload path in case padding bytes were
initialized on GPU. The workaround is not desirable to keep. Change the
buffer padding zero initialization code to initialize padding bytes on
CPU if buffer is host visible which avoids the need to transition usage.

This also requires changing NonzeroClearResourcesOnCreationForTesting
path for host visible buffers since that must happen before zero
initializing the padding bytes.

UploadData() is also modified to only zero initialize data if memcpy()
isn't going to write to the full buffer. This matches the behavior in
EnsureDataInitializedAsDestination() when buffer is initialized on GPU.

Bug: 460180744
Change-Id: I1ada9e30919ee3acb4c3c7ebbbfe13d0c41f005c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/275134
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>