[dawn] Minor fix for sizeof on bitfield
Looks like when dawn assert are disabled we fall down into a
macro for DAWN_ASSERT that uses sizeof(conditional).

This for logical reasons is not allowed.

I not could find an easy fix to Assert.h so lets fix this for now
this way.

Change-Id: Ibd901bcb1e44b7f041455c9776bf5107f858ae20
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/276474
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>