[bindless] Setup core and wgsl def entries for resource table.
Add the `getResource` and `hasResource` entires to the core.def and
wgsl.def files. Setup the backends to reject the new methods in the
CanGenerate methods.

The `getResource` auto-generated tests are skipped as they break the
template (they need to be used in a texture function, not stored like
all other cases are).

Fixed: 463344960
Change-Id: If9b4beb545fbfe09d0f4ab9e884d79376026bc69
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/276376
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>