Add a tiled matrix multiply test to Dawn
This is a more interesting test of subgroup matrix than existing
coverage.

It is also used to hit an issue with the Metal backend where the MSL
compiler produces a pipeline that has a maximum number of threads per
threadgroup that is lower than the workgroup size specified in the
WGSL shader.

Bug: 443794633
Change-Id: I8850cd1069901058cd763f121f0c9088486cde6e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/272175
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: David Neto <dneto@google.com>