[uniformity] fix connectivity of for-loop without condition
In a loop, a barrier followed by a non-uniform break or return
should cause a uniformity failure.

When a for loop has no condition expression, the control flow node
for the start of the loop body did not connect to the
'start of iteration' control flow node.

The fix is to initialize the start-of-body node to be the
start-of-iter node.

The change makes the following code fail, where before it did not.

fn main(@builtin(local_invocation_index) nonuniform_var: u32, @builtin(num_workgroups) nw: vec3u) {
  let uniform_cond = 42 == nw.x;
  let nonuniform_cond = 42 == nonuniform_var;
  for ( ; ; ) {
    workgroupBarrier();
    if nonuniform_cond { break; }
  }
}

Bug: 463704875
Change-Id: Ieca2905ed2842195cebbd5930e6048316a6a6964
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/276755
Commit-Queue: David Neto <dneto@google.com>
Reviewed-by: James Price <jrprice@google.com>