Vulkan: Only use `ALLOW_VARYING_SUBGROUP_SIZE` when `SubgroupSizeControl` isn't enabled
This patch only use
`VK_PIPELINE_SHADER_STAGE_CREATE_ALLOW_VARYING_SUBGROUP_SIZE_BIT`
when the Vulkan extension `SubgroupSizeControl` is not enabled
because according to Vulkan SPEC that flag cannot be set if a
`VkPipelineShaderStageRequiredSubgroupSizeCreateInfo` structure is
included in the pNext chain.

This patch also removes the toggle `VulkanDisallowVaryingSubgroupSize`
as we no longer need it now.

Bug: 410595996, 365910276
Change-Id: I65b520f3124e9c4344e60d1110138d0b20289ed3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/275594
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Shao, Jiawei <jiawei.shao@intel.com>