Rename split FSTestOSWrapper files
Renames the FSTestOSWrapper test files that have been split out so far
to not reference the component they are testing. It turns out that
FSTestOSWrapper implements most things on the FilesystemReaderWriter
component instead of having FilesystemReader and FilsystemWriter.
Additionally, this distinction is likely going away in the future
anyways once we make it CWD-aware, which will require accessing the
environment.

Bug: 436025865
Change-Id: Ic2659648ebba6a8fbdd600056825601de2c8ceca
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/276819
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@google.com>