Remove TSAN suppression
Removes a suppression for a TSAN failure in SharedTextureMemoryTests
that was recently added. It turns out that the issue being caught is a
false positive and should be suppressed at the TSAN level instead of
skipping the test. This can also occur flakily in other tests, so
per-test suppressions are not feasible.

Bug: 462477486
Change-Id: I0c30709a9bbdd2116ebaa61c51caf5e31858e6a6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/277695
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@google.com>
Auto-Submit: Brian Sheedy <bsheedy@google.com>