[dawn][native] Lazily initialize dynamic array resources
This CL sets up all the machinery to make BindGroup.Update() and friends
work and update the descriptors in the underlying VkDescriptorSets.
It tracks dirty resources in the DynamicArrayState and applies resource
updates in the Vulkan backend at the same time as the metadata buffer
updates.

The initialization code for dynamic array BindGroups is no longer needed
in the Vulkan backend as initial bindings are set using `Update`
internally, so InitializeDynamicBindings is turned into a function to
apply the ResourceUpdates instead.

Bug: 435317394
Change-Id: I8210add0198c2220e063b7fec7838f00011376ff
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/278136
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>