Refactor DawnPerfTestWithParams SetUp
DawnPerfTestWithParams::SetUp has a DAWN_TEST_UNSUPPORTED_IF
call inside which is supposed to skip the test. But, it was
called from derived SetUp functions which would then continue
to run a bunch of setup code, even though all of that was
invalid and supposed to be skipped.

Refactored so DawnPerfTestWithParams has a SetUpPerfTest
virtual function that is only called if
DawnPerfTestWithParams::SetUp succeeds.

Change-Id: Iec1a5c23d5bf6cb6970a481c133c68266a6a6964
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/278255
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Shrek Shao <shrekshao@google.com>
Commit-Queue: Gregg Tavares <gman@chromium.org>