Remove autolock from Buffer::Unmap()
Don't lock the device mutex when wgpu::Buffer::Unmap() is called. This
primarily relies on access to buffers being externally synchronized by
the dawn client.

Buffer MapAsync() or Unmap() shouldn't overlap with other use of the
buffer but if it does a validation error will be thrown. Buffer
Destroy() is similar but once BufferBase::DestroyImpl() runs there is no
way to abort destroying the buffer. In this case DestroyImpl() will spin
loop until the concurrent operation is done before destroying.

To guard against concurrent use a new BufferState is added to denote the
buffer is inside an operation. At the start of MapAsync() and Unmap()
the buffer state is changed with compare_exchange() and any concurrent
changes will cause compare_exchange() to fail so it can be handled.

The backend may still need to lock the device mutex during UnmapImpl()
if it's not thread safe. For Vulkan and Metal that isn't required so
unmapping a buffer won't block.

Note that removing device mutex from MapAsync() and Destroy() will
happen as a follow up.

Bug: 425472913
Change-Id: I79f8314e8bcbf8a1eb5672cf28e72edb984e7643
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262535
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>
Reviewed-by: Loko Kung <lokokung@google.com>