Disable DXGI's alt+enter fullscreen handling
DXGI will listen for alt+enter and go fullscreen in response to this shortcut, sometimes in
broken ways. Disable this behavior to allow applications to handle this shortcut using their
own logic (if desired). Note that MakeWindowAssociation must be called after swapchain
creation, and using the same DXGIFactory used to create the swapchain.