Run test spec presubmit on DEPS changes
Refactors the existing presubmit to ensure that the generated JSON files
under //infra/specs are up to date and runs it on changes to DEPS. This
helps catch issues of //testing being updated through DEPS and the test
specs getting out of date due to a change in Swarming dimensions.

Since a gclient sync is necessary to pull in the updated //testing
contents, it is possible for a manual roll to be uploaded without this
check being hit locally. However, the Dawn presubmit bot should fail in
such cases.

Bug: 452840620
Change-Id: Iaadf3a53b58c8bbd6e8c4b402a1ec80dc01b8415
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/281375
Commit-Queue: Brian Sheedy <bsheedy@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>