[dawn][native] Disallow 0 texture usage even with internal usages
Previously `usage` could be zero as long as `internalUsage` was nonzero.
But the spec requires `usage=0` to be a validation error. The extension
shouldn't change this.

This fixes a test I'm adding in https://github.com/gpuweb/cts/pull/4527
(It doesn't affect the public `createTexture()` because that can't add
internal usages, but it does affect *canvas* textures.)

Fixed: 468353736
Change-Id: Ibeb6d4b539701ee93d6039792e62d5a77aa2ca71
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/279775
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>